requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
python-dateutil==2.9.0.post0
requests==2.32.3
urllib3==2.2.3
python-dotenv==1.0.1
